Add autodeaf so users in other groups see status
This causes the user to be self_mute and self_deaf when there
is no SIP call active. Users in other Mumble groups can then
see whether the user is active without moving to same group.

Note: this needs the updated mumlib with the self_mute and self_deaf
methods